+++ sdk/identity/azure_identity/src/credentials/azure_cli_credentials.rs
@@ -3,6 +3,8 @@
 
 use crate::{
     credentials::{cache::TokenCache, TokenCredentialOptions},
+    env::Env,
+    process::{shell_exec, OutputProcessor},
     validate_scope, validate_subscription, validate_tenant_id,
 };
 use azure_core::{
@@ -12,7 +14,7 @@ use azure_core::{
     process::{new_executor, Executor},
 };
 use serde::Deserialize;
-use std::{ffi::OsStr, fmt::Debug, str, sync::Arc};
+use std::{ffi::OsString, sync::Arc};
 use time::OffsetDateTime;
 use tracing::trace;
 
@@ -42,6 +44,27 @@ impl CliTokenResponse {
                 "expires_on field not found. Please use Azure CLI 2.54.0 or newer.",
             )),
         }
+    }
+}
+
+impl OutputProcessor for CliTokenResponse {
+    fn credential_name() -> &'static str {
+        "AzureCliCredential"
+    }
+
+    fn get_error_message(_stderr: &str) -> Option<&str> {
+        // Azure CLI's errors are generally clear and more helpful than anything we'd write here
+        None
+    }
+
+    fn deserialize_token(stdout: &str) -> azure_core::Result<AccessToken> {
+        let response: Self = from_json(stdout)?;
+        let expires_on = response.expires_on()?;
+        Ok(AccessToken::new(response.access_token, expires_on))
+    }
+
+    fn tool_name() -> &'static str {
+        "az"
     }
 }
 
@@ -76,6 +99,8 @@ pub struct AzureCliCredentialOptions {
     /// If `None`, one is created using [`new_executor`]; alternatively,
     /// you can supply your own implementation using a different asynchronous runtime.
     pub executor: Option<Arc<dyn Executor>>,
+
+    env: Option<Env>,
 }
 
 impl AzureCliCredential {
@@ -88,6 +113,9 @@ impl AzureCliCredential {
         if let Some(ref subscription) = options.subscription {
             validate_subscription(subscription)?;
         }
+        if options.env.is_none() {
+            options.env = Some(Env::default());
+        }
         if options.executor.is_none() {
             options.executor = Some(new_executor());
         }
@@ -98,89 +126,36 @@ impl AzureCliCredential {
         }))
     }
 
-    /// Get an access token for an optional resource
-    async fn get_access_token(&self, scopes: &[&str]) -> azure_core::Result<CliTokenResponse> {
+    async fn get_token(&self, scopes: &[&str]) -> azure_core::Result<AccessToken> {
         if scopes.is_empty() {
             return Err(Error::new(
                 ErrorKind::Credential,
                 "exactly one scope required",
             ));
         }
-        // Pass the CLI a Microsoft Entra ID v1 resource because we don't know which CLI version is installed and older ones don't support v2 scopes.
-        let resource = scopes[0].trim_end_matches("/.default");
-        validate_scope(resource)?;
+        validate_scope(scopes[0])?;
 
-        // On Windows az is a cmd and it should be called like this.
-        // See https://doc.rust-lang.org/nightly/std/process/struct.Command.html
-        let program = if cfg!(target_os = "windows") {
-            "cmd"
-        } else {
-            "az"
-        };
-        let mut args = Vec::new();
-        if cfg!(target_os = "windows") {
-            args.push("/C");
-            args.push("az");
-        }
-        args.push("account");
-        args.push("get-access-token");
-        args.push("--output");
-        args.push("json");
-        args.push("--resource");
-        args.push(resource);
-
+        let mut command = OsString::from("az account get-access-token -o json --scope ");
+        command.push(scopes[0]);
         if let Some(ref tenant_id) = self.options.tenant_id {
-            args.push("--tenant");
-            args.push(tenant_id);
+            command.push(" --tenant ");
+            command.push(tenant_id);
         }
         if let Some(ref subscription) = self.options.subscription {
-            args.push("--subscription");
-            args.push(subscription);
+            command.push(r#" --subscription ""#);
+            command.push(subscription);
+            command.push("\"");
         }
 
-        trace!(
-            "fetching credential via Azure CLI: {program} {}",
-            args.join(" "),
-        );
+        trace!("running Azure CLI command: {command:?}");
 
-        let args = args.iter().map(|arg| arg.as_ref()).collect::<Vec<&OsStr>>();
-
-        let status = self
-            .options
-            .executor
-            .as_ref()
-            // It's okay to call unwrap() here because new() ensures it's initialized.
-            .unwrap()
-            .run(OsStr::new(program), &args)
-            .await;
-        match status {
-            Ok(az_output) if az_output.status.success() => {
-                let output = str::from_utf8(&az_output.stdout)?;
-
-                let access_token = from_json(output)?;
-                Ok(access_token)
-            }
-            Ok(az_output) => {
-                let output = String::from_utf8_lossy(&az_output.stderr);
-                Err(Error::with_message(ErrorKind::Credential, || {
-                    format!("'az account get-access-token' command failed: {output}")
-                }))
-            }
-            Err(e) => match e.kind() {
-                std::io::ErrorKind::NotFound => {
-                    Err(Error::message(ErrorKind::Other, "Azure CLI not installed"))
-                }
-                error_kind => Err(Error::with_message(ErrorKind::Other, || {
-                    format!("Unknown error of kind: {error_kind:?}")
-                })),
-            },
-        }
-    }
-
-    async fn get_token(&self, scopes: &[&str]) -> azure_core::Result<AccessToken> {
-        let tr = self.get_access_token(scopes).await?;
-        let expires_on = tr.expires_on()?;
-        Ok(AccessToken::new(tr.access_token, expires_on))
+        shell_exec::<CliTokenResponse>(
+            // unwrap() is safe because new() ensured the values are Some
+            self.options.executor.clone().unwrap(),
+            self.options.env.as_ref().unwrap(),
+            &command,
+        )
+        .await
     }
 }
 
@@ -204,22 +179,166 @@ impl From<TokenCredentialOptions> for AzureCliCredentialOptions {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::tests::{MockExecutor, FAKE_TENANT_ID, FAKE_TOKEN, LIVE_TEST_SCOPES};
+    use std::ffi::OsStr;
+    use time::{format_description::well_known::Rfc3339, UtcOffset};
 
-    #[test]
-    fn read_cli_token_response() -> azure_core::Result<()> {
-        let json = br#"
-        {
-            "accessToken": "MuchLonger_NotTheRealOne_Sv8Orn0Wq0OaXuQEg",
-            "expiresOn": "2024-01-01 19:23:16.000000",
-            "expires_on": 1704158596,
-            "subscription": "33b83be5-faf7-42ea-a712-320a5f9dd111",
-            "tenant": "065e9f5e-870d-4ed1-af2b-1b58092353f3",
-            "tokenType": "Bearer"
-        }
-        "#;
-        let token_response: CliTokenResponse = from_json(json)?;
-        assert_eq!(token_response.expires_on, Some(1704158596));
-        assert_eq!(token_response.expires_on()?.unix_timestamp(), 1704158596);
-        Ok(())
+    async fn run_test(
+        exit_code: i32,
+        stdout: &str,
+        stderr: &str,
+        subscription: Option<String>,
+        tenant_id: Option<String>,
+    ) -> azure_core::Result<AccessToken> {
+        let subscription_for_on_run = subscription.clone();
+        let tenant_for_on_run = tenant_id.clone();
+        let system_root = "/dev/null";
+        let options =
+            AzureCliCredentialOptions {
+                env: Some(Env::from(&[("SYSTEMROOT", system_root)][..])),
+                executor: Some(MockExecutor::with_output(
+                    exit_code,
+                    stdout,
+                    stderr,
+                    Some(Arc::new(move |program: &OsStr, args: &[&OsStr]| {
+                        let args: Vec<String> = args
+                            .iter()
+                            .map(|arg| arg.to_string_lossy().to_string())
+                            .collect();
+                        if cfg!(target_os = "windows") {
+                            assert_eq!(program.to_string_lossy(), "cmd");
+                            assert_eq!(args[0], "/C");
+                            assert!(args[1].starts_with(&format!(
+                                "cd {system_root} && az account get-access-token -o json"
+                            )));
+                        } else {
+                            assert_eq!(program, "/bin/sh");
+                            assert_eq!(args[0], "-c");
+                            assert!(args[1]
+                                .starts_with("cd /bin && az account get-access-token -o json"));
+                        }
+                        for scope in LIVE_TEST_SCOPES {
+                            assert!(args[1].contains(&format!(" --scope {scope}")));
+                        }
+                        if let Some(ref subscription_id) = subscription_for_on_run {
+                            assert!(args[1]
+                                .contains(&format!(r#" --subscription "{subscription_id}""#)));
+                        } else {
+                            assert!(!args[1].contains("--subscription"));
+                        }
+                        if let Some(ref tenant_id) = tenant_for_on_run {
+                            assert!(args[1].contains(&format!(" --tenant {tenant_id}")));
+                        } else {
+                            assert!(!args[1].contains("--tenant"));
+                        }
+                    })),
+                )),
+                tenant_id,
+                subscription,
+                ..Default::default()
+            };
+        let cred = AzureCliCredential::new(Some(options))?;
+        return cred.get_token(LIVE_TEST_SCOPES).await;
+    }
+
+    #[tokio::test]
+    async fn error_includes_stderr() {
+        let stderr = "something went wrong";
+        let err = run_test(1, "stdout", stderr, None, None)
+            .await
+            .expect_err("expected error");
+        assert!(matches!(err.kind(), ErrorKind::Credential));
+        assert!(err.to_string().contains(stderr));
+    }
+
+    #[tokio::test]
+    async fn get_token_success() {
+        let expires_on = OffsetDateTime::parse("2038-01-18T00:00:00Z", &Rfc3339).unwrap();
+        let stdout = format!(
+            r#"{{"accessToken":"{FAKE_TOKEN}",
+            "expiresOn":"2030-01-02 03:04:05.000000",
+            "expires_on":{},
+            "subscription":"...",
+            "tenant":"{FAKE_TENANT_ID}",
+            "tokenType":"Bearer"}}"#,
+            expires_on.unix_timestamp()
+        );
+
+        let token = run_test(0, &stdout, "", None, None).await.expect("token");
+
+        assert_eq!(FAKE_TOKEN, token.token.secret());
+        assert_eq!(UtcOffset::UTC, token.expires_on.offset());
+        assert_eq!(expires_on, token.expires_on);
+    }
+
+    #[tokio::test]
+    async fn not_logged_in() {
+        let err = run_test(1, "", "Please run 'az login' to setup account.", None, None)
+            .await
+            .expect_err("error");
+        assert!(matches!(err.kind(), ErrorKind::Credential));
+        assert!(err.to_string().contains("az login"));
+    }
+
+    #[tokio::test]
+    async fn program_not_found() {
+        let executor = MockExecutor::with_error(std::io::Error::from_raw_os_error(127));
+        let options = AzureCliCredentialOptions {
+            executor: Some(executor),
+            ..Default::default()
+        };
+
+        let err = AzureCliCredential::new(Some(options))
+            .expect("valid credential")
+            .get_token(LIVE_TEST_SCOPES)
+            .await
+            .expect_err("expected error");
+
+        assert!(matches!(err.kind(), ErrorKind::Credential));
+    }
+
+    #[tokio::test]
+    async fn subscription() {
+        let expires_on = OffsetDateTime::parse("2038-01-18T00:00:00Z", &Rfc3339).unwrap();
+        let subscription = "subscription name";
+        let stdout = format!(
+            r#"{{"accessToken":"{FAKE_TOKEN}",
+            "expiresOn":"2030-01-02 03:04:05.000000",
+            "expires_on":{},
+            "subscription":"{subscription}",
+            "tenant":"{FAKE_TENANT_ID}",
+            "tokenType":"Bearer"}}"#,
+            expires_on.unix_timestamp()
+        );
+
+        let token = run_test(0, &stdout, "", Some(subscription.to_string()), None)
+            .await
+            .expect("token");
+
+        assert_eq!(FAKE_TOKEN, token.token.secret());
+        assert_eq!(UtcOffset::UTC, token.expires_on.offset());
+        assert_eq!(expires_on, token.expires_on);
+    }
+
+    #[tokio::test]
+    async fn tenant_id() {
+        let expires_on = OffsetDateTime::parse("2038-01-18T00:00:00Z", &Rfc3339).unwrap();
+        let stdout = format!(
+            r#"{{"accessToken":"{FAKE_TOKEN}",
+            "expiresOn":"2030-01-02 03:04:05.000000",
+            "expires_on":{},
+            "subscription":"...",
+            "tenant":"{FAKE_TENANT_ID}",
+            "tokenType":"Bearer"}}"#,
+            expires_on.unix_timestamp()
+        );
+
+        let token = run_test(0, &stdout, "", None, Some(FAKE_TENANT_ID.to_string()))
+            .await
+            .expect("token");
+
+        assert_eq!(FAKE_TOKEN, token.token.secret());
+        assert_eq!(UtcOffset::UTC, token.expires_on.offset());
+        assert_eq!(expires_on, token.expires_on);
     }
 }